+++ node-manager/src/error.rs
@@ -1,70 +1,79 @@
-use gloo_storage::errors::StorageError;
-use std::fmt;
+use thiserror::Error;
 
-#[derive(Debug)]
+#[derive(Error, Debug)]
 #[allow(dead_code)]
 // copied from LDK lite
 /// An error that possibly needs to be handled by the user.
-pub enum Error {
+pub enum MutinyError {
     /// Returned when trying to start Mutiny while it is already running.
+    #[error("Mutiny is already running.")]
     AlreadyRunning,
     /// Returned when trying to stop Mutiny while it is not running.
+    #[error("Mutiny is not running.")]
     NotRunning,
     /// The funding transaction could not be created.
+    #[error("Funding transaction could not be created.")]
     FundingTxCreationFailed,
     /// A network connection has been closed.
+    #[error("Network connection closed.")]
     ConnectionFailed,
     /// Payment of the given invoice has already been initiated.
+    #[error("An invoice must not get payed twice.")]
     NonUniquePaymentHash,
     /// The given invoice is invalid.
+    #[error("The given invoice is invalid.")]
     InvoiceInvalid,
     /// Invoice creation failed.
+    #[error("Failed to create invoice.")]
     InvoiceCreationFailed,
     /// No route for the given target could be found.
+    #[error("Failed to find route.")]
     RoutingFailed,
     /// A given peer info could not be parsed.
+    #[error("Failed to parse the given peer information.")]
     PeerInfoParseFailed,
     /// A channel could not be opened.
+    #[error("Failed to create channel.")]
     ChannelCreationFailed,
     /// A channel could not be closed.
+    #[error("Failed to close channel.")]
     ChannelClosingFailed,
     /// Persistence failed.
-    PersistenceFailed,
+    #[error("Failed to persist data.")]
+    PersistenceFailed {
+        #[from]
+        source: MutinyStorageError,
+    },
     /// A wallet operation failed.
+    #[error("Failed to conduct wallet operation.")]
     WalletOperationFailed,
     /// A signing operation failed.
+    #[error("Failed to sign given transaction.")]
     WalletSigningFailed,
     /// A chain access operation failed.
+    #[error("Failed to conduct chain access operation.")]
     ChainAccessFailed,
+    #[error(transparent)]
+    Other(#[from] anyhow::Error),
 }
 
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Self::AlreadyRunning => write!(f, "Mutiny is already running."),
-            Self::NotRunning => write!(f, "Mutiny is not running."),
-            Self::FundingTxCreationFailed => {
-                write!(f, "Funding transaction could not be created.")
-            }
-            Self::ConnectionFailed => write!(f, "Network connection closed."),
-            Self::NonUniquePaymentHash => write!(f, "An invoice must not get payed twice."),
-            Self::InvoiceInvalid => write!(f, "The given invoice is invalid."),
-            Self::InvoiceCreationFailed => write!(f, "Failed to create invoice."),
-            Self::RoutingFailed => write!(f, "Failed to find route."),
-            Self::PeerInfoParseFailed => write!(f, "Failed to parse the given peer information."),
-            Self::ChannelCreationFailed => write!(f, "Failed to create channel."),
-            Self::ChannelClosingFailed => write!(f, "Failed to close channel."),
-            Self::PersistenceFailed => write!(f, "Failed to persist data."),
-            Self::WalletOperationFailed => write!(f, "Failed to conduct wallet operation."),
-            Self::WalletSigningFailed => write!(f, "Failed to sign given transaction."),
-            Self::ChainAccessFailed => write!(f, "Failed to conduct chain access operation."),
-        }
-    }
+#[derive(Error, Debug)]
+pub enum MutinyStorageError {
+    #[error("Failed to use browser storage")]
+    StorageError {
+        #[from]
+        source: gloo_storage::errors::StorageError,
+    },
+    #[error("Failed to serialize or deserialize")]
+    SerdeError {
+        #[from]
+        source: serde_json::Error,
+    },
+    #[error(transparent)]
+    Other(#[from] anyhow::Error),
 }
 
-impl std::error::Error for Error {}
-
-impl From<bdk::Error> for Error {
+impl From<bdk::Error> for MutinyError {
     fn from(e: bdk::Error) -> Self {
         match e {
             bdk::Error::Signer(_) => Self::WalletSigningFailed,
@@ -73,15 +82,16 @@ impl From<bdk::Error> for Error {
     }
 }
 
-// todo uncomment when we add esplora stuff
-// impl From<esplora::EsploraError> for Error {
-//     fn from(_e: esplora::EsploraError) -> Self {
-//         Self::ChainAccessFailed
-//     }
-// }
-
-impl From<StorageError> for Error {
-    fn from(_e: StorageError) -> Self {
-        Self::PersistenceFailed
+impl From<MutinyStorageError> for bdk::Error {
+    fn from(e: MutinyStorageError) -> Self {
+        match e {
+            MutinyStorageError::StorageError { source } => {
+                bdk::Error::Generic(format!("Storage error: {source}"))
+            }
+            MutinyStorageError::SerdeError { source } => {
+                bdk::Error::Generic(format!("Serde error: {source}"))
+            }
+            _ => bdk::Error::Generic("Unexpected Mutiny storage Error".to_string()),
+        }
     }
 }
